Refactor: move createNewCardList to Board folder
