+++ .eslintrc.json
@@ -24,6 +24,7 @@
     "no-unused-vars": "error",
     "n/handle-callback-err": "off",
     "no-multiple-empty-lines": "error",
+    "@typescript-eslint/no-confusing-void-expression": "off",
     "@typescript-eslint/no-floating-promises": "off",
     "@typescript-eslint/promise-function-async": "off",
     "@typescript-eslint/consistent-type-imports": "off",
