+++ src/queries/cards/index.ts
@@ -3,8 +3,9 @@ import { AxiosError } from "axios";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { SEARCH_PARAMS_KEY } from "@/constants";
 import { request } from "@/utils/httpRequest";
+import { createNewCardList } from "@components/Board/createNewCardList";
 import { RequestParams } from "@/interfaces/httpRequest";
-import { ICard, ICardList } from "@/interfaces/cards";
+import { ICardList } from "@/interfaces/cards";
 import * as I from "./interface";
 
 const cardListsKeys = {
@@ -131,51 +132,6 @@ export const useEditCardPositionMutation = () => {
       },
     },
   );
-};
-
-const createNewCardList = (currentCardList: ICardList[], cardId: string, listId: string, index: number) => {
-  const currentCardsList = currentCardList.find((cardList) => cardList.cards.some((card) => card.id === cardId));
-  const currentListId = currentCardsList?.id;
-  const card = currentCardsList?.cards.find((card) => card.id === cardId);
-  if (!card) return;
-
-  const updatedCardLists = currentCardList.map((cardList) => {
-    const isDroppedList = cardList.id === listId;
-    const isDraggedList = cardList.id === currentListId;
-    const isMovedToSameList = isDroppedList && isDraggedList;
-
-    if (isMovedToSameList) {
-      return reorderCardInList(cardList, card, cardId, index);
-    }
-
-    if (isDroppedList) {
-      return addCardToList(cardList, card, index);
-    }
-
-    if (isDraggedList) {
-      return removeCardFromList(cardList, cardId);
-    }
-
-    return cardList;
-  });
-
-  return updatedCardLists;
-};
-
-const reorderCardInList = (cardList: ICardList, card: ICard, cardId: string, index: number) => {
-  const filteredCardList = removeCardFromList(cardList, cardId);
-  return addCardToList(filteredCardList, card, index);
-};
-
-const addCardToList = (cardList: ICardList, card: ICard, index: number) => {
-  const newCards = Array.from(cardList.cards);
-  newCards.splice(index, 0, card);
-  return { ...cardList, cards: newCards };
-};
-
-const removeCardFromList = (cardList: ICardList, cardId: string) => {
-  const filteredCards = cardList.cards.filter((card) => card.id !== cardId);
-  return { ...cardList, cards: filteredCards };
 };
 
 export const useDeleteListMutation = () => {
